Remove the previous baseurl
With the repository moved to the Landskrona SK organization, the
repository is configured to be the root GitHub Page for the
organization, hence the baseurl is no longer necessary.

This is what will allow us to configure the CNAME so that we can get our
old landskrona.schack.se back working again; previously that didn't work
since it becamse landskrona.schack.se/LandskronaSK, i.e. with this
baseurl which was the name of the repository in my own private account.